@@ -4,9 +4,9 @@ version = "0.1.0"
 description = "FastAPI-based REST API for managing clinical and genetic data for individuals with HNF1B disease"
 requires-python = ">=3.10"
 dependencies = [
+    # Core FastAPI dependencies
     "email-validator>=2.3.0",
     "fastapi>=0.116.1",
-    "motor>=3.7.1",
     "pandas>=2.3.2",
     "passlib>=1.7.4",
     "pydantic>=2.11.7",
@@ -15,42 +15,22 @@ dependencies = [
     "python-dotenv>=1.1.1",
     "python-multipart>=0.0.20",
     "uvicorn[standard]>=0.35.0",
+    
+    # PostgreSQL and database dependencies
+    "sqlalchemy[asyncio]>=2.0.25",
+    "asyncpg>=0.29.0",
+    "alembic>=1.13.1",
+    "psycopg2-binary>=2.9.9",
+    
+    # Scientific dependencies (for sheets import)
+    "biopython>=1.84",
+    "requests>=2.31.0",
 ]
 
-[tool.black]
-line-length = 88
-target-version = ['py38']
-include = '\.pyi?$'
-extend-exclude = '''
-/(
-  # directories
-  \.eggs
-  | \.git
-  | \.hg
-  | \.mypy_cache
-  | \.tox
-  | \.venv
-  | build
-  | dist
-  | data
-)/
-'''
-
-[tool.isort]
-profile = "black"
-line_length = 88
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
 
 [dependency-groups]
 dev = [
-    "black==23.12.1",
-    "flake8==7.0.0",
-    "flake8-docstrings==1.7.0",
-    "isort==5.13.2",
+    "ruff==0.6.8",
     "mypy==1.8.0",
 ]
 test = [
@@ -68,5 +48,34 @@ no_implicit_optional = true
 explicit_package_bases = true
 
 [[tool.mypy.overrides]]
-module = ["motor.*", "pandas.*", "passlib.*"]
+module = ["pandas.*", "passlib.*", "biopython.*", "Bio.*"]
 ignore_missing_imports = true
+
+[tool.ruff]
+line-length = 88
+target-version = "py310"
+
+[tool.ruff.lint]
+select = [
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "F",  # pyflakes
+    "I",  # isort
+    "D",  # pydocstyle
+]
+ignore = [
+    "D100",  # Missing docstring in public module
+    "D104",  # Missing docstring in public package
+    "D203",  # 1 blank line required before class docstring
+    "D213",  # Multi-line docstring summary should start at the second line
+]
+
+[tool.ruff.lint.pydocstyle]
+convention = "google"
+
+[tool.ruff.lint.isort]
+# Configuration compatible with black formatting
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
