@@ -1,13 +1,7 @@
 # File: app/endpoints/genes.py
-import time
 from typing import Any, Dict, Optional
 
-from bson import ObjectId
 from fastapi import APIRouter, HTTPException, Query, Request
-from fastapi.encoders import jsonable_encoder
-
-from app.database import db
-from app.utils import build_pagination_meta, parse_filters, parse_sort
 
 router = APIRouter()
 
@@ -25,57 +19,7 @@ async def get_genes(
         ),
     ),
 ) -> Dict[str, Any]:
-    """
-    Retrieve a paginated list of genes.
-
-    This endpoint returns gene structure data stored in the 'genes' collection.
-    It supports JSON:API–style filtering via query parameters, for example:
-      /genes?filter[status]=active&sort=-gene_symbol&page=2&page_size=10
-    """
-    start_time = time.perf_counter()  # Start timing
-
-    # Build filters from query parameters.
-    query_params = dict(request.query_params)
-    filters = parse_filters(query_params)
-
-    # Determine the sort option (default is ascending by "gene_symbol").
-    sort_option = parse_sort(sort) if sort else ("gene_symbol", 1)
-
-    collection = db.genes
-    total = await collection.count_documents(filters)
-    skip_count = (page - 1) * page_size
-
-    # Build the MongoDB cursor using skip() and limit().
-    cursor = collection.find(filters)
-    if sort_option:
-        cursor = cursor.sort(*sort_option)
-    cursor = cursor.skip(skip_count).limit(page_size)
-    genes = await cursor.to_list(length=page_size)
-
-    if not genes:
-        raise HTTPException(status_code=404, detail="No genes found")
-
-    base_url = str(request.url).split("?")[0]
-    end_time = time.perf_counter()  # End timing
-    exec_time = end_time - start_time
-
-    # Build extra query parameters (retain all except pagination-specific ones).
-    extra_params: Dict[str, Any] = {
-        k: v for k, v in request.query_params.items() if k not in {"page", "page_size"}
-    }
-
-    # Build pagination metadata including execution time.
-    meta = build_pagination_meta(
-        base_url,
-        page,
-        page_size,
-        total,
-        query_params=extra_params,
-        execution_time=exec_time,
+    """Retrieve a paginated list of genes with structure data."""
+    raise HTTPException(
+        status_code=503, detail="Genes endpoint temporarily unavailable"
     )
-
-    # Convert the result to JSON-friendly data (e.g., convert ObjectId to str).
-    response_data = jsonable_encoder(
-        {"data": genes, "meta": meta}, custom_encoder={ObjectId: lambda o: str(o)}
-    )
-    return response_data
