@@ -1,13 +1,7 @@
 # File: app/endpoints/proteins.py
-import time
 from typing import Any, Dict, Optional
 
-from bson import ObjectId
 from fastapi import APIRouter, HTTPException, Query, Request
-from fastapi.encoders import jsonable_encoder
-
-from app.database import db
-from app.utils import build_pagination_meta, parse_filters, parse_sort
 
 router = APIRouter()
 
@@ -25,58 +19,7 @@ async def get_proteins(
         ),
     ),
 ) -> Dict[str, Any]:
-    """
-    Retrieve a paginated list of proteins.
-
-    This endpoint queries the `proteins` collection for protein structure and
-    domain data.
-    It supports JSON:API–style filtering via query parameters.
-    For example:
-      /proteins?filter[domain]=kinase&sort=-gene&page=2&page_size=10
-    """
-    start_time = time.perf_counter()  # Start timing
-
-    # Build filters from the request query parameters.
-    query_params = dict(request.query_params)
-    filters = parse_filters(query_params)
-
-    # Use provided sort or default to sorting by "gene" ascending.
-    sort_option = parse_sort(sort) if sort else ("gene", 1)
-
-    collection = db.proteins
-    total = await collection.count_documents(filters)
-    skip_count = (page - 1) * page_size
-
-    cursor = collection.find(filters)
-    if sort_option:
-        cursor = cursor.sort(*sort_option)
-    cursor = cursor.skip(skip_count).limit(page_size)
-    proteins = await cursor.to_list(length=page_size)
-
-    if not proteins:
-        raise HTTPException(status_code=404, detail="No proteins found")
-
-    base_url = str(request.url).split("?")[0]
-    end_time = time.perf_counter()  # End timing
-    execution_time = end_time - start_time
-
-    # Build extra query parameters by retaining all except pagination-specific ones.
-    extra_params: Dict[str, Any] = {
-        k: v for k, v in request.query_params.items() if k not in {"page", "page_size"}
-    }
-
-    # Build pagination metadata including execution time.
-    meta = build_pagination_meta(
-        base_url,
-        page,
-        page_size,
-        total,
-        query_params=extra_params,
-        execution_time=execution_time,
+    """Retrieve a paginated list of proteins with structure and domain data."""
+    raise HTTPException(
+        status_code=503, detail="Protein endpoint temporarily unavailable"
     )
-
-    # Convert the proteins and metadata into JSON-friendly types.
-    response_data = jsonable_encoder(
-        {"data": proteins, "meta": meta}, custom_encoder={ObjectId: lambda o: str(o)}
-    )
-    return response_data
